+++ src/v/kafka/server/connection_context.h
@@ -114,7 +114,7 @@ private:
     // used to track number of pending requests
     class request_tracker {
     public:
-        explicit request_tracker(rpc::server_probe& probe) noexcept
+        explicit request_tracker(net::server_probe& probe) noexcept
           : _probe(probe) {
             _probe.request_received();
         }
@@ -126,7 +126,7 @@ private:
         ~request_tracker() noexcept { _probe.request_completed(); }
 
     private:
-        rpc::server_probe& _probe;
+        net::server_probe& _probe;
     };
     // used to pass around some internal state
     struct session_resources {
